kuis_service/requirements.txt
Flask==3.0.0
graphene==3.3
graphql-server[flask]==3.0.0b7
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==0.19.0 